requirements.txt
dropbox~=11.36.0
watchdog~=3.0.0
pysmb~=1.2.9.1
smbprotocol~=1.10.1
paramiko~=3.1.0